# lib/api-client-react/src/generated/api.schemas.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API
 * OpenAPI spec version: 0.1.0
 */
export interface HealthStatus {
  status: string;
}

export interface MessageResponse {
  message: string;
}

export type UserRole = typeof UserRole[keyof typeof UserRole];


export const UserRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface User {
  id: string;
  clerkId: string;
  email: string;
  name: string;
  /** @nullable */
  avatarUrl?: string | null;
  role: UserRole;
  isApproved: boolean;
  createdAt: string;
}

export interface UserList {
  users: User[];
  total: number;
}

export interface UserProfileUpdate {
  name?: string;
}

export type UserAdminUpdateRole = typeof UserAdminUpdateRole[keyof typeof UserAdminUpdateRole];


export const UserAdminUpdateRole = {
  admin: 'admin',
  user: 'user',
} as const;

export interface UserAdminUpdate {
  role?: UserAdminUpdateRole;
  isApproved?: boolean;
}

export interface Invite {
  id: string;
  code: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  note?: string | null;
  createdBy: string;
  /** @nullable */
  usedBy?: string | null;
  /** @nullable */
  usedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}

export interface InviteInput {
  email?: string;
  note?: string;
  expiresAt?: string;
}

export type ContentItemType = typeof ContentItemType[keyof typeof ContentItemType];


export const ContentItemType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface ContentItem {
  id: number;
  tmdbId: number;
  type: ContentItemType;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  overview: string;
  /** @nullable */
  releaseDate?: string | null;
  rating: number;
  voteCount: number;
  genres: string[];
}

export type ContentDetailType = typeof ContentDetailType[keyof typeof ContentDetailType];


export const ContentDetailType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface SeasonInfo {
  seasonNumber: number;
  name: string;
  episodeCount: number;
  /** @nullable */
  posterPath?: string | null;
}

export interface CastMember {
  id: number;
  name: string;
  character: string;
  /** @nullable */
  profilePath?: string | null;
}

export interface ContentDetail {
  id: number;
  tmdbId: number;
  type: ContentDetailType;
  title: string;
  /** @nullable */
  posterPath?: string | null;
  /** @nullable */
  backdropPath?: string | null;
  overview: string;
  /** @nullable */
  releaseDate?: string | null;
  rating: number;
  voteCount: number;
  genres: string[];
  /** @nullable */
  runtime?: number | null;
  /** @nullable */
  status?: string | null;
  /** @nullable */
  tagline?: string | null;
  seasons?: SeasonInfo[];
  cast: CastMember[];
  /** @nullable */
  trailerKey?: string | null;
}

export interface EmbedUrlResponse {
  embedUrl: string;
  provider: string;
}

export interface SearchResults {
  results: ContentItem[];
  totalResults: number;
  page: number;
  totalPages: number;
}

export type RoomContentType = typeof RoomContentType[keyof typeof RoomContentType];


export const RoomContentType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface Room {
  id: string;
  name: string;
  contentType: RoomContentType;
  contentId: string;
  contentTitle: string;
  /** @nullable */
  contentPoster?: string | null;
  hostId: string;
  hostName: string;
  memberCount: number;
  currentTime: number;
  isPlaying: boolean;
  /** @nullable */
  season?: number | null;
  /** @nullable */
  episode?: number | null;
  createdAt: string;
}

export type RoomInputContentType = typeof RoomInputContentType[keyof typeof RoomInputContentType];


export const RoomInputContentType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export interface RoomInput {
  name: string;
  contentType: RoomInputContentType;
  contentId: string;
  contentTitle: string;
  contentPoster?: string;
  season?: number;
  episode?: number;
}

export type ChatMessageType = typeof ChatMessageType[keyof typeof ChatMessageType];


export const ChatMessageType = {
  text: 'text',
  reaction: 'reaction',
  system: 'system',
} as const;

export interface ChatMessage {
  id: string;
  roomId: string;
  userId: string;
  userName: string;
  /** @nullable */
  userAvatar?: string | null;
  content: string;
  type: ChatMessageType;
  createdAt: string;
}

export type ListUsersParams = {
page?: number;
limit?: number;
};

export type GetTrendingParams = {
type?: GetTrendingType;
};

export type GetTrendingType = typeof GetTrendingType[keyof typeof GetTrendingType];


export const GetTrendingType = {
  all: 'all',
  movie: 'movie',
  tv: 'tv',
} as const;

export type GetPopularParams = {
type?: GetPopularType;
};

export type GetPopularType = typeof GetPopularType[keyof typeof GetPopularType];


export const GetPopularType = {
  movie: 'movie',
  tv: 'tv',
} as const;

export type SearchContentParams = {
q: string;
type?: SearchContentType;
page?: number;
};

export type SearchContentType = typeof SearchContentType[keyof typeof SearchContentType];


export const SearchContentType = {
  all: 'all',
  movie: 'movie',
  tv: 'tv',
} as const;

export type GetEmbedUrlParams = {
type: GetEmbedUrlType;
id: string;
season?: number;
episode?: number;
};

export type GetEmbedUrlType = typeof GetEmbedUrlType[keyof typeof GetEmbedUrlType];


export const GetEmbedUrlType = {
  movie: 'movie',
  tv: 'tv',
} as const;