# lib/api-zod/src/generated/types/invite.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API
 * OpenAPI spec version: 0.1.0
 */

export interface Invite {
  id: string;
  code: string;
  /** @nullable */
  email?: string | null;
  /** @nullable */
  note?: string | null;
  createdBy: string;
  /** @nullable */
  usedBy?: string | null;
  /** @nullable */
  usedAt?: string | null;
  /** @nullable */
  expiresAt?: string | null;
  createdAt: string;
}