# lib/api-zod/src/generated/types/seasonInfo.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * StreamVault API
 * OpenAPI spec version: 0.1.0
 */

export interface SeasonInfo {
  seasonNumber: number;
  name: string;
  episodeCount: number;
  /** @nullable */
  posterPath?: string | null;
}